requirements.txt
# AI and Bot
pysocks>=1.7.1  # For SOCKS5 proxy support
python-telegram-bot>=20.0
google-generativeai>=0.3.0
chromadb>=0.4.0
pillow>=10.0.0
# Web Scraping
requests>=2.31.0
httpx[socks]>=0.26.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
parsel>=1.8.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Browser automation (for JS-rendered content)
playwright>=1.40.0

# Data processing
pandas>=2.1.0
pydantic>=2.5.0
python-slugify>=8.0.1

# CLI
click>=8.1.0
rich>=13.0.0

# Utils
python-dotenv>=1.0.0
tenacity>=8.2.0
tqdm>=4.66.0

# Export
openpyxl>=3.1.0  # Excel export

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=23.12.0
ruff>=0.1.0
